diarization/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
pyannote.audio==3.1.1
torch==2.1.2
torchaudio==2.1.2
pydub==0.25.1
numpy<2.0.0
scipy>=1.11.0
huggingface_hub<1.0.0
python-dotenv==1.0.1
